Migrated from python subprocess to QT QProcess
The calling of programs was migrated to QProcess instead of using python subprocess.
Doing this means one less import, and that buttons can now be disabled while the
program is running so that users don't click a zillion times waiting for a slow program
to load.

The readme was tweaked slightly in hopes of pleasing github's rst2html script regarding pre
blocks.