Config and stylesheet checks
When the configured stylesheet cannot be found, a warning is given to stdout and the default QT style is used.  Previously an exception was thrown.

The program searches for a config file in various directories, so it doesn't have to be explicitly provided.  The last value found of: /etc/kilauncher.yaml, ~/.kilauncher.yaml, or the argument of -c will be the config file.

If one isn't found, an error is printed to stdout and the program exits.